Rearranged in MVC architecture
Huge change in main.py file because most of it went to views.py file.
