+++ website/templates/navbar.html
@@ -1,5 +1,8 @@
 <nav class="navbar fixed-top p-3  navbar-expand-lg navbar-dark bg-dark">
   <div class="container-fluid">
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
     <span class="navbar-brand">Life Pharma</span>
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav me-auto mb-2 mb-lg-0">
@@ -10,10 +13,10 @@
           <a class="nav-link" href="#">About</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="{{ url_for('search_page') }}">Search</a>
+          <a class="nav-link" href="{{ url_for('views.search_page') }}">Search</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="{{ url_for('camera_page') }}">Camera</a>
+          <a class="nav-link" href="{{ url_for('views.camera_page') }}">Camera</a>
         </li>
       </ul>
     </div>
